# server/src/main/resources/open-api-contract.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenFireStationManager API",
...
    "contact": {
      "name": "OpenFireStationManager",
      "url": "https://github.com/simonhauck/OpenFireStationManager"
    },
    "version": "v1"
  },
  "security": [ {
    "sessionCookie": [ ]
  } ],
...
  "paths": {
...
    "/api/clothing/types": {
      "get": {
        "tags": [ "Clothing Types" ],
        "summary": "List every clothing type",
        "description": "Returns all garment categories (`Kleidungsart`) known to the station, such as `Einsatzjacke` or `Helm`. Types are the vocabulary the rest of the clothing API is built on: start here to discover the valid `typeId` values before creating or searching for items.\n\nThe list is small and is returned unpaginated.\n\n**Authorisation:** any signed-in user.",
        "operationId": "listClothingTypes",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClothingType"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated — no valid session cookie was supplied.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error. The response body carries no details.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [ "Clothing Types" ],
        "summary": "Create a new clothing type",
        "description": "Registers a new garment category. Do this before creating items of a kind the station has never held before.\n\nNames are not enforced to be unique, so check the existing list first to avoid creating a near-duplicate that will split the overview reports in two.\n\n**Authorisation:** requires the `KLEIDERWART` role. An `ADMIN` account implicitly holds every role and may also call this.",
        "operationId": "createClothingType",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrUpdateClothingTypeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The newly created type.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ClothingType"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated — no valid session cookie was supplied.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but the account lacks the required role.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error. The response body carries no details.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "400": {
            "description": "The request failed validation. The `errors` array names each offending field.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetail"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/clothing/types/{id}": {
      "get": {
        "tags": [ "Clothing Types" ],
        "summary": "Get one clothing type by its numeric id",
        "description": "Looks up a single garment category by its database id.\n\n**Authorisation:** any signed-in user.",
        "operationId": "getClothingType",
        "parameters": [ {
          "name": "id",
          "in": "path",
          "description": "Numeric id of the clothing type.",
          "required": true,
          "schema": {
            "type": "integer",
            "format": "int64",
            "exclusiveMinimum": 0
          },
          "example": 3
        } ],
        "responses": {
          "200": {
            "description": "The clothing type.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ClothingType"
                }
              }
            }
          },
          "404": {
            "description": "No clothing type exists with this id.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated — no valid session cookie was supplied.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error. The response body carries no details.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "400": {
            "description": "The request failed validation. The `errors` array names each offending field.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetail"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [ "Clothing Types" ],
        "summary": "Delete a clothing type",
        "description": "Removes a garment category. This is only possible once no clothing item references the type any more — the request is refused with `409 Conflict` otherwise, rather than cascading. Delete or re-type the remaining items first.\n\n**Authorisation:** requires the `KLEIDERWART` role. An `ADMIN` account implicitly holds every role and may also call this.",
        "operationId": "deleteClothingType",
        "parameters": [ {
          "name": "id",
          "in": "path",
          "description": "Numeric id of the clothing type to delete.",
          "required": true,
          "schema": {
            "type": "integer",
            "format": "int64",
            "exclusiveMinimum": 0
          },
          "example": 3
        } ],
        "responses": {
          "204": {
            "description": "The type was deleted."
          },
          "404": {
            "description": "No clothing type exists with this id.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "Clothing items still reference this type. Nothing was deleted; the message states how many items remain.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated — no valid session cookie was supplied.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but the account lacks the required role.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error. The response body carries no details.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "400": {
            "description": "The request failed validation. The `errors` array names each offending field.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetail"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [ "Clothing Types" ],
        "summary": "Rename a clothing type",
        "description": "Changes the display name of a category. Items already assigned to this type keep their assignment and immediately report the new name.\n\n**Authorisation:** requires the `KLEIDERWART` role. An `ADMIN` account implicitly holds every role and may also call this.",
        "operationId": "updateClothingType",
        "parameters": [ {
          "name": "id",
          "in": "path",
          "description": "Numeric id of the clothing type to rename.",
          "required": true,
          "schema": {
            "type": "integer",
            "format": "int64",
            "exclusiveMinimum": 0
          },
          "example": 3
        } ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrUpdateClothingTypeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The updated type.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ClothingType"
                }
              }
            }
          },
          "404": {
            "description": "No clothing type exists with this id.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated — no valid session cookie was supplied.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but the account lacks the required role.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error. The response body carries no details.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "400": {
            "description": "The request failed validation. The `errors` array names each offending field.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetail"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ProblemDetail": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "format": "uri"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string",
            "format": "uri"
          },
          "properties": {
            "type": "object",
            "additionalProperties": { }
          }
        }
      },
      "ValidationError": {
        "description": "A single field-level validation failure.",
        "properties": {
          "field": {
            "type": "string",
            "description": "Path of the offending field within the request body, e.g. `size`. Empty when the rule applies to the request as a whole rather than one field.",
            "example": "size"
          },
          "message": {
            "type": "string",
            "description": "Human-readable reason the value was rejected.",
            "example": "must not be blank"
          },
          "rejectedValue": {
            "type": "object",
            "description": "The value that was submitted and rejected. May be null."
          }
        }
      },
      "ValidationProblemDetail": {
        "description": "RFC 9457 problem document describing a request that failed validation, extended with an `errors` array naming each offending field.",
        "properties": {
          "type": {
            "type": "string",
            "example": "about:blank"
          },
          "title": {
            "type": "string",
            "example": "Invalid request"
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "example": 400
          },
          "detail": {
            "type": "string",
            "example": "Validation failed for request body"
          },
          "instance": {
            "type": "string",
            "description": "URI of the request that failed, when available."
          },
          "errors": {
            "type": "array",
            "description": "Every field-level failure detected in this request.",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
...
      "EntityMetaData": {
        "type": "object",
        "description": "Audit trail attached to every stored record. Maintained by the server on save — these fields are ignored if supplied in a request. `System` appears as the actor for changes made outside a user session, such as database migrations.",
        "properties": {
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "When the record was first stored."
          },
          "createdBy": {
            "type": "string",
            "description": "Username of whoever created the record, or `System`.",
            "example": "m.mustermann"
          },
          "lastModifiedAt": {
            "type": "string",
            "format": "date-time",
            "description": "When the record was last changed."
          },
          "lastModifiedBy": {
            "type": "string",
            "description": "Username of whoever last changed the record, or `System`.",
            "example": "m.mustermann"
          }
        },
        "required": [ "createdAt", "createdBy", "lastModifiedAt", "lastModifiedBy" ]
      },
...
      "CreateOrUpdateClothingTypeRequest": {
        "type": "object",
        "description": "The name of a clothing type. Used for both creation and update.",
        "properties": {
          "name": {
            "type": "string",
            "description": "Display name of the category, e.g. `Einsatzjacke` or `Helm`. Names are not required to be unique, but duplicates make the overview reports ambiguous.",
            "example": "Einsatzjacke",
            "maxLength": 255,
            "minLength": 0
          }
        },
        "required": [ "name" ]
      },
      "ClothingType": {
        "type": "object",
        "description": "A category of garment (`Kleidungsart`), such as `Einsatzjacke` or `Helm`. Types carry no size or quantity of their own — those live on the individual clothing items.",
        "properties": {
          "name": {
            "type": "string",
            "description": "Display name of the category.",
            "example": "Einsatzjacke"
          },
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "Server-assigned identifier.",
            "example": 3
          },
          "metaData": {
            "$ref": "#/components/schemas/EntityMetaData"
          }
        },
        "required": [ "id", "metaData", "name" ]
      },
...
    },
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "description": "Server-side session cookie issued by `POST /api/public/auth/login`. Send it on every authenticated request.",
        "name": "OFSM_AUTH",
        "in": "cookie"
      },
      "rememberMeCookie": {
        "type": "apiKey",
        "description": "Long-lived cookie issued when logging in with `rememberMe = true`, valid for 30 days. It authenticates on its own, so a long-lived client may send it instead of a session cookie.",
        "name": "OFSM_AUTH_REMEMBER_ME",
        "in": "cookie"
      }
    }
  }
}